refactor(bank): Remove useless "Enum" in enumerations names
